Bugfix: Always use the received package name when its PackageInfo exist
This fixes a bug where the installing status is stalled due to wrongly using the old package name, when the newer one is received. Instead, always use the package name if it exists in the repository.

Let rPkgName be the received package name from broadcast instead, and oPkgName be its other equivalent based on samePackagesMap