add support for fs-verity
RPackages can now have an optional "hasFsVeritySignatures" property.
If it's true and the device trusts one of the repo certificates,
fs-verity signatures (fsv_sig) are downloaded and written into the installer session.

Installation with fsv_sig will complete successfully only if fs-verity was enabled for all APKs
of the package, which will happen only if fsv_sig comes from the certificate that is included in the
system image.