bugfix: inconsistent release channel to package mapping
In collectDependencies(), preferred release channel was required to match exactly with package
release channel, otherwise most stable package variant was returned.

Elsewhere, preferred release channel was treated as a lower bound of package stability instead of
an exact match requirement.

In practice, this meant that when preferred release channel was set to alpha, but only beta and
stable packages were present in the repo, stable package was picked instead of beta package in some
places, leading to an incorrect dependency resolution process.