maintain and use a global list of busy packages
Package is considered busy if it's being installed in any user by Apps app.
Package is marked as busy when installation starts and is unmarked when installation completes.
If package is already busy when installation starts, installation is aborted.

Error message about this is shown if the installation was started manually and is skipped if it was
started automatically (eg by auto-update job).