respect "showAutoUpdateNotification" package property in "Updates available" notification
Previously, it was used only when deciding whether to show the post-auto-update notification.

On GrapheneOS, when auto-update is disabled, this will stop "Update available" notifications for
GmsCompatConfig and will list only Vanadium in this notification instead of Vanadium, WebView,
Trichrome Library (this will also show available Vanadium version in notification if it's the only
available update, since available version is not shown when there's more than one available update).