declare DEFAULT and APP_MARKET intent category
To add the Apps app to the default home screen for new setups, the
Launcher app needs to identify it as an app market app. This is done by
allowing implicit intents with the DEFAULT intent category and adding
the APP_MARKET intent category, enabling the Launcher app to recognize
and include it on the home screen.