switch foreground service type to special use
Data sync has a limit of 6 hours within a 24 hour period in Android 15.
The 6 hour limit is highly unlikely to be reached for our App Store in a
real world use case but could be hit in a contrived case such as running
tests of the app all day trying to find bugs.

Special use needs to be approved for the Play Store, but that doesn't
impact our App Store. We can't publish this app through the Play Store
regardless and it wouldn't make sense since it should be handling
updating itself. If we wanted to make it obtainable through the Play
Store, we'd likely want to do it with a tiny bootstrap app which then
asks to remove itself after installing the App Store so it can take over
updating itself.